pyproject.toml
[tool.poetry]
name = "online-store"
version = "0.1.0"
description = ""
authors = ["kooznitsa <jukuznetswork@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.3"
gunicorn = "^21.2.0"
uvicorn = "^0.28.1"
aiohttp = "^3.9.3"
python-dotenv = "^1.0.1"
pillow = "^10.2.0"
psycopg2-binary = "^2.9.9"
celery = "^5.3.6"
flower = "^2.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
